requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==24.2.0
audioop-lts==0.2.1
beautifulsoup4==4.12.3
chardet==4.0.0
discord==2.3.2
discord.py==1.7.3
frozenlist==1.5.0
idna==3.10
legacy-cgi==2.6.1
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
soupsieve==2.6
typing_extensions==4.12.2
yarl==1.18.0
